+++ changes.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v1.0.2] - 2025-08-14
+### Changed
+- Minor update to `server.js` and `index.html` adding "exclude BIN" and "reinclude BIN" option to status dropdown.
+- Minor update to `server.js` and `index.html` adding function to restore curator decisions to a new .db file from changes.log
+- Minor update to `server.js` and `index.html` adding version number of server.js and index.html at the bottom of the user interface.
+- Minor update to `user_manual_windows.html` explaining the new functions.
+
+--> in order to get all function, replace `server.js`, `index.html` and `user_manual_windows.html` with the new version (see user_manual_windows.html, chapter 9)!
+
 ## [v1.0.1] - 2025-07-24
 ### Changed
 - Minor update to `server.js` to improve dataset loading performance.